pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "information-crawler"
version = "0.2.0"
description = "Information monitoring system for AI research academy"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.0",
    "httpx>=0.28",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "feedparser>=6.0",
    "apscheduler>=3.10,<4.0",
    "playwright>=1.49",
    "pyyaml>=6.0",
    "scalar-fastapi>=1.0",
    "supabase>=2.0.0",
    "tqdm>=4.66",
    "asyncpg>=0.29",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
